Fixed engine_start_region_search_event and created engine_region_loaded_event and tested it
Fixed start_region_search_event where it returns None if the region does not exist. The region load event will load a selected region's information.

This function searches for regions given the region_code, local_code, and/or name. When there are multiple results it returns a list of the results which will be processed in main.