Changed conditional check for engine_save_edited_country and engine_save_new_region and added try-except to engine_save_edited_region.
For the conditionals for engine_save_edited_country and engine_save_new_region, I changed it so the conditional checks for values from the view such as None and 0 rather than '' to clean up which error is raised. In addition, the try-except in save_edited_region checks if user tries to change to value already in the database while allowing them to keep the same region_code value if they want.